Version 0.1.8 Build 1
NEW:
1. Added temporary background music and several sound effects to the
game.
2. Health bar now actually show the remaining player health.

CHANGES:
1. Changed player movement. Now Ruzik’s plane will move the same
distance as the user’s finger, but with a limited speed (so it may take
Ruzik longer to move to the desired position than it took the user to
move his finger). This also makes his movements seem a little unstable…
Will continue working on it.

FIXES:
1. Second weapons (currently defaults to missiles) firing button now
works normally and is completely separate from player movements.